Changed types used for big integers to handle differences between Linux and OSX
* Added types seq32_t and seq64_t for sequence numbers
* Using (u)llint_t defined as long long for all big numbers that must be
  minimum 64 bits to handle different definitions on Linux and OSX.

+++ time_util.h
@@ -16,8 +16,8 @@ char* sprint_time_sec_prec(char *buf, struct timeval t);
 char* sprint_readable_time_now(char *buf);
 struct timeval sprint_readable_time_diff(char *buf, struct timeval oldest, struct timeval newest);
 struct timeval sprint_readable_time_now_diff(char *buf, struct timeval old_time);
-int get_msecs(struct timeval tv);
-int get_usecs(timeval &tv);
+long get_msecs(struct timeval tv);
+long get_usecs(timeval &tv);
 
 #define tval_pair(tval) tval.tv_sec, tval.tv_usec
 
@@ -27,19 +27,19 @@ int get_usecs(timeval &tv);
 	do {															\
 		(result)->tv_sec = (a)->tv_sec - (b)->tv_sec;				\
 		(result)->tv_usec = (a)->tv_usec - (b)->tv_usec;			\
-		if ( (result)->tv_sec > 0) {								\
+		if ((result)->tv_sec > 0) {									\
 			if ((result)->tv_usec < 0) {							\
 				--(result)->tv_sec;									\
 				(result)->tv_usec += 1000000;						\
 			}														\
-		} else if ( (result)->tv_sec < 0 ) {						\
+		} else if ((result)->tv_sec < 0) {							\
 			if ((result)->tv_usec > 0) {							\
 				++(result)->tv_sec;									\
 				(result)->tv_usec = 1000000 - (result)->tv_usec;	\
 			} else { /* if (tv_usec < 0) */							\
 				(result)->tv_usec *= -1;							\
 			}														\
-			if((result)->tv_sec == 0 )								\
+			if ((result)->tv_sec == 0)								\
 				(result)->tv_usec *= -1;							\
 		}															\
 	} while (0)
